requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Flask-Script==2.0.6
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
paramiko==2.7.2
protobuf==3.16.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
pysmb==1.2.6
pyspnego==0.1.6
pytz==2021.1
six==1.16.0
smbprotocol==1.5.1
Werkzeug==0.16.1
